refactor: common function
