+++ pkg/attackflow/aws_api_gateway.go
@@ -2,7 +2,6 @@ package attackflow
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"strings"
@@ -120,36 +119,6 @@ func handleAPIGatewayError(err error) error {
 		return nil
 	}
 	return err
-}
-
-func parseMetadata(metadata interface{}) (string, error) {
-	metaJSON, err := json.Marshal(metadata)
-	if err != nil {
-		return "", err
-	}
-	return string(metaJSON), nil
-}
-
-func setNode(isPublic bool, resource *datasource.Resource, resp *datasource.AnalyzeAttackFlowResponse) *datasource.AnalyzeAttackFlowResponse {
-	if isPublic {
-		internet := getInternetNode()
-		if !existsInternetNode(resp.Nodes) {
-			resp.Nodes = append(resp.Nodes, internet)
-		}
-		resp.Edges = append(resp.Edges, getEdge(internet.ResourceName, resource.ResourceName, ""))
-	}
-	resp.Nodes = append(resp.Nodes, resource)
-	return resp
-}
-
-func getExternalServiceNode(target string) *datasource.Resource {
-	return &datasource.Resource{
-		ResourceName: target,
-		ShortName:    "external-service",
-		Layer:        LAYER_EXTERNAL_SERVICE,
-		Region:       REGION_GLOBAL,
-		Service:      "external-service",
-	}
 }
 
 func extractApiID(arn string) (string, error) {
