+++ pkg/attackflow/aws_api_gateway_v1.go
@@ -84,6 +84,6 @@ func (a *apiGatewayAnalyzer) analyzeV1(ctx context.Context, resp *datasource.Ana
 	if err != nil {
 		return nil, err
 	}
-	resp = setNode(a.metadata.IsPublic, a.resource, resp)
+	resp = setNode(a.metadata.IsPublic, "api", a.resource, resp)
 	return resp, nil
 }
